refactor: cleaned up project structure
