+++ cast-fe/src/components/index.js
@@ -1,27 +1,21 @@
-import Chat from "./Chat";
-import Dashboard from "./Dashboard";
-import Fresh from "./Fresh";
-import Home from "./Home";
-import Scene from "./Scene";
-import Search from "./Search";
-import Subscribed from "./Subscribed";
-import Trending from "./Trending";
-import Liked from "./Liked";
-import Live from "./Live";
-import Manage from "./Manage";
-import Profile from "./Profile";
+import Author from "./Author";
+import Cast from "./Cast";
+import CastEditable from "./CastEditable";
+import HybridPlayer from "./HybridPlayer";
+import List from "./List";
+import Navigation from "./Navigation";
+import Sidebar from "./Sidebar";
+import SidebarProfile from "./SidebarProfile";
+import Toast from "./Toast";
 
 export {
-  Chat,
-  Dashboard,
-  Fresh,
-  Home,
-  Liked,
-  Live,
-  Manage,
-  Profile,
-  Scene,
-  Search,
-  Subscribed,
-  Trending,
+  Author,
+  Cast,
+  CastEditable,
+  HybridPlayer,
+  List,
+  Navigation,
+  Sidebar,
+  SidebarProfile,
+  Toast,
 };
