--- cast-fe/src/components/player/index.js
@@ -1,3 +1,0 @@
-import HybridPlayer from "./HybridPlayer";
-
-export { HybridPlayer };
