+++ cast-fe/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 import "bootstrap/dist/css/bootstrap.min.css";
-import "./index.css";
+import "./styles/index.css";
 
 ReactDOM.render(<App />, document.getElementById("root"));
