+++ cast-fe/src/styles/player.css
