+++ cast-fe/src/views/Profile.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Card, Col, Container, Row } from "react-bootstrap";
-import SidebarProfile from "./SidebarProfile";
+import { SidebarProfile } from "../components";
 import auth from "../helper/auth";
 import MediaQuery from "react-responsive";
 import { MOBILE_BP } from "../constants/breakpoint";
