+++ cast-fe/src/views/index.js
@@ -1,0 +1,27 @@
+import Chat from "./Chat";
+import Dashboard from "./Dashboard";
+import Fresh from "./Fresh";
+import Home from "./Home";
+import Liked from "./Liked";
+import Live from "./Live";
+import Manage from "./Manage";
+import Profile from "./Profile";
+import Scene from "./Scene";
+import Search from "./Search";
+import Subscribed from "./Subscribed";
+import Trending from "./Trending";
+
+export {
+  Chat,
+  Dashboard,
+  Fresh,
+  Home,
+  Liked,
+  Live,
+  Manage,
+  Profile,
+  Scene,
+  Search,
+  Subscribed,
+  Trending,
+};
